### 2020-12-26 #2931 @danielskatz
👋 @bauer-alex - note that whedon thinks some of your references might be missing DOIs - if it is correct, please add them, then use the commands
* ```@whedon generate PDF```
* ```@whedon check references```
(each at the start of a new comment) to check again and regenerate the PDF
