### 2020-12-26 #2931 @danielskatz
Note that this is a new version of https://doi.org/10.21105/joss.00557 (from about 3 years ago), and should be checked to make sure it has enough new content for JOSS to re-review it
