### 2020-12-26 #2931 @danielskatz
👋 @mbobra - would you be willing to edit this submission for JOSS? (and I know that you are out through 4 Jan, so I'm not expecting an immediate response)
