### 2020-12-26 8ebe963
JOSS: added missing doi's
