+++ paper/paper.bib
@@ -6,7 +6,8 @@
   volume={105},
   pages={46--52},
   year={2017},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.csda.2016.07.010}
 }
 
 @Manual{goldsmith2016,
@@ -17,49 +18,6 @@
   url = {https://github.com/jeff-goldsmith/gfpca},
 }
 
-@Manual{refund.shiny,
-   title = {refund.shiny: Interactive plotting for functional data analyses},
-   author = {Wrobel, J and Goldsmith, J},
-   year = {2015},
-   note = {R package version 0.2},
- }
- 
-
-@article{rcpp,
-   title = {{Rcpp}: Seamless {R} and {C++} Integration},
-   author = {Dirk Eddelbuettel and Romain Fran\c{c}ois},
-   journal = {Journal of Statistical Software},
-   year = {2011},
-   volume = {40},
-   number = {8},
-   pages = {1--18},
-   url = {http://www.jstatsoft.org/v40/i08/},
-   doi = {10.18637/jss.v040.i08},
-}
-  
-@article{rcppArma, 
-   title = {RcppArmadillo: Accelerating R with high-performance C++ linear algebra},
-   author = {Dirk Eddelbuettel and Conrad Sanderson},
-   journal = {Computational Statistics and Data Analysis},
-   year = {2014},
-   volume = {71},
-   month = {March},
-   pages = {1054--1063},
-   url = {http://dx.doi.org/10.1016/j.csda.2013.02.005},
-}
-  
-  
-  
-@article{wrobel2016,
-  Author = {Wrobel, J and Park, S-Y and Staicu, A-M and Goldsmith, J},
-  Journal = {Stat},
-  title = {Interactive graphics for functional data analyses},
-  volume = {5},
-  pages = {88-101},
-  year = {2016},
-  doi = {10.1002/sta4.109}
-}
-
 @article{wrobel2018,
   title={register: Registration for Exponential Family Functional Data},
   author={Wrobel, J},
@@ -67,7 +25,8 @@
   volume={3},
   number={22},
   pages={557},
-  year={2018}
+  year={2018},
+  doi={10.21105/joss.00557}
 }
 
 @article{wrobel2019,
@@ -78,5 +37,6 @@
   number={1},
   pages={48--57},
   year={2019},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1111/biom.12963}
 }
