### 2020-12-26 #2931 @bauer-alex
All missing doi's are now included. Thanks for the hint!
