### 2020-12-27 70efffd
minor code cleanup in cov_hall
