### 2021-01-05 4e42487
adjusted penalty for full incompleteness, also in JOSS submission
