### 2021-01-05 74fdd9c
JOSS minor fix in the formula
