### 2021-01-08 797207c
two-step GFPCA: renamed 'relevant digits' to 'significant digits'
