+++ tests/testthat/test-gfpca_twoStep.R
@@ -1,23 +1,23 @@
 context("gfpca_twoStep")
 
-test_that("coarsen_index only accepts positive integer 'relevant_digits'",{
+test_that("coarsen_index only accepts positive integer 'significant_digits'",{
 	index_values = 1:3
 	
-	expect_that(coarsen_index(index_values, relevant_digits = 0),
-							throws_error("'relevant_digits' must be a positive integer."))
+	expect_that(coarsen_index(index_values, significant_digits = 0),
+							throws_error("'significant_digits' must be a positive integer."))
 })
 
 test_that("coarsen_index correctly rounds positive values",{
 	index_values1 = c(0.84729, 0.9379, 0.19328)
 	index_values2 = c(307, 86938, 13)
 	
-	expect_equal(coarsen_index(index_values1, relevant_digits = 1),
+	expect_equal(coarsen_index(index_values1, significant_digits = 1),
 							 expected = c(0.8, 0.9, 0.2))
-	expect_equal(coarsen_index(index_values1, relevant_digits = 3),
+	expect_equal(coarsen_index(index_values1, significant_digits = 3),
 							 expected = c(0.847, 0.938, 0.193))
-	expect_equal(coarsen_index(index_values2, relevant_digits = 1),
+	expect_equal(coarsen_index(index_values2, significant_digits = 1),
 							 expected = c(0, 90000, 0))
-	expect_equal(coarsen_index(index_values2, relevant_digits = 4),
+	expect_equal(coarsen_index(index_values2, significant_digits = 4),
 							 expected = c(310, 86940, 10))
 })
 
@@ -25,13 +25,13 @@ test_that("coarsen_index correctly rounds negative values",{
 	index_values1 = c(0.84729, -0.9379, 0.19328)
 	index_values2 = c(-307, 86938, -13)
 	
-	expect_equal(coarsen_index(index_values1, relevant_digits = 1),
+	expect_equal(coarsen_index(index_values1, significant_digits = 1),
 							 expected = c(0.8, -0.9, 0.2))
-	expect_equal(coarsen_index(index_values1, relevant_digits = 3),
+	expect_equal(coarsen_index(index_values1, significant_digits = 3),
 							 expected = c(0.847, -0.938, 0.193))
-	expect_equal(coarsen_index(index_values2, relevant_digits = 1),
+	expect_equal(coarsen_index(index_values2, significant_digits = 1),
 							 expected = c(0, 90000, 0))
-	expect_equal(coarsen_index(index_values2, relevant_digits = 4),
+	expect_equal(coarsen_index(index_values2, significant_digits = 4),
 							 expected = c(-310, 86940, -10))
 })
 
@@ -65,7 +65,7 @@ test_that("gfpca_twoStep (Gaussian) output is a list with non-null items and cla
 test_that("gfpca_twoStep (Gaussian) output has correct number of dimensions",{
 	Y = simulate_functional_data(I = 100, D = 200, seed = 2020)$Y
 	Kt = 8
-	fpca_object = gfpca_twoStep(Y, npc = 2, Kt = Kt, index_relevantDigits = 4)
+	fpca_object = gfpca_twoStep(Y, npc = 2, Kt = Kt, index_significantDigits = 4)
 	
 	expect_equal(length(fpca_object$t_vec), 200)
 	expect_equal(dim(fpca_object$alpha), c(200, 1))
@@ -74,7 +74,7 @@ test_that("gfpca_twoStep (Gaussian) output has correct number of dimensions",{
[…]

[…]
