### 2021-01-15 #2964 @whedon
[REVIEW]: registr 2.0: Incomplete Curve Registration for Exponential Family Functional Data
**Submitting author:** @bauer-alex (<a href="http://orcid.org/0000-0003-3495-5131">Alexander Bauer</a>)
**Repository:** <a href="https://github.com/julia-wrobel/registr" target ="_blank">https://github.com/julia-wrobel/registr</a>
**Version:** v2.1.0
