### 2021-01-16 2c5e74b
register_fpca: return detailed convergence information
