### 2021-01-16 0047601
minor update of JOSS paper and internal 'growth_incomplete' data
