+++ tests/testthat/test-registr.R
@@ -174,7 +174,7 @@ test_that("registr function only accepts Y_template if it has the correct format
 	Y_template1   = Y[Y$id %in% template_ids1,]
 	template_ids2 = "girl12"
 	Y_template2   = Y[Y$id %in% template_ids2,]
-	template_ids3 = c("boy01","boy04","boy29","boy30","boy31","boy34","boy36")
+	template_ids3 = c("girl12","boy04","boy29")
 	Y_template3   = Y[Y$id %in% template_ids3,]
 	
 	expect_error(registr(Y = Y, family = "gaussian", Y_template = Y_template1$value),
