+++ man/growth_incomplete.Rd
@@ -23,9 +23,9 @@ first derivative, the curves were slightly smoothed. \cr \cr
 To showcase the functionality of the \code{registr} package regarding the
 analysis of incomplete curves, the growth curves were artificially made
 incomplete. For each child, leading incompleteness was simulated by drawing
-a random initial age in the first half of the domain (i.e., ages [1.25,8.5]).
+a random initial age in the first quarter of the domain.
 Also, trailing incompleteness was simulated by drawing
-a random cut-off age in the second half of the domain (i.e., ages [9.5,18]).
+a random cut-off age in the second half of the domain.
 }
 \references{
 Ramsay, J. O., and Silverman, B. W. (2006),
