### 2021-01-16 #2964 @bauer-alex
I made a minor update of the paper, so I'm updating the pdf.
