### 2021-01-19 f58076d
two-step FPCA: minor change of handling of negative eigenvalues
