### 2021-01-23 c674b2f
registr and register_fpca now return all elements to reconstruct the full estimated warping functions
