### 2021-01-24 1d7d8a7
minor fix of Yhat matrix for the two-step GFPCA
