+++ R/gfpca_twoStep.R
@@ -239,7 +239,7 @@ gfpca_twoStep = function (Y, family = "gaussian", npc = 1, Kt = 8,
   
   ## format output
   fittedVals = data.frame(
-    id    = rep(1:I, each = D),
+    id    = rep(ids, each = D),
     index = rep(output_index, I),
     value = as.vector(t(Yhat))
   )
