### 2021-01-25 1bcbc73
register_fpca: Changed convergence threshold from 0.00001 to 0.0001
