+++ R/register_fpca.R
@@ -60,6 +60,8 @@
 #' Alexander Bauer \email{alexander.bauer@@stat.uni-muenchen.de}
 #' @export
 #' 
+#' @importFrom magrittr %>%
+#' 
 #' @return An object of class \code{registration} containing:
 #' \item{Y}{The observed data plus variables \code{t_star} and \code{t_hat} which are the
 #' unregistered grid and registered grid, respectively.}
@@ -86,7 +88,7 @@
 #' Vector of mean squared differences between the (warped) index values
 #' (scaled to [0,1] based on the size of the observed domain)
 #' in the current and the previous iteration.
-#' Convergence is reached if this measure drops below 0.00001. \cr \cr
+#' Convergence is reached if this measure drops below 0.0001. \cr \cr
 #' \emph{registration_loss} \cr
 #' Vector of the loss in each iteration of the algorithm.
 #' Calculated in the registration step using the exponential family
@@ -192,7 +194,7 @@ register_fpca = function(Y, Kt = 8, Kh = 4, family = "gaussian",
   
   iter        = 0
   delta_index = rep(NA, max_iterations)
-  convergence_threshold = 0.00001
+  convergence_threshold = 0.0001
   
   while (iter == 0 ||
   			 (iter < max_iterations && delta_index[iter] > convergence_threshold)) {
