### 2021-02-02 6bd1079
added github workflow for checks
