+++ R/plot.fpca.R
@@ -25,8 +25,6 @@
 #' @param xlab,ylab Optional titles for the x and y axis.
 #' @param ... Additional arguments passed to \code{\link[ggplot2]{theme}}.
 #' 
-#' @import ggplot2
-#' @importFrom cowplot plot_grid
 #' @importFrom dplyr bind_rows
 #' @export
 #' 
@@ -38,13 +36,20 @@
 #' data(growth_incomplete)
 #' 
 #' fpca_obj = fpca_gauss(Y = growth_incomplete, npc = 2)
+#' if (requireNamespace("ggplot2", quietly = TRUE) &&
+#' requireNamespace("cowplot", quietly = TRUE)) {
 #' plot(fpca_obj)
+#' }
 #' 
 plot.fpca = function(x, plot_npc = x$npc, var_factor = 2,
                      response_function = NULL,
                      subtitle = TRUE, xlim = NULL, ylim = NULL,
                      xlab = "t [registered]", ylab = "y", ...) {
   
+  if (!requireNamespace("ggplot2", quietly = TRUE) || 
+      !requireNamespace("cowplot", quietly = TRUE)) {
+    stop("ggplot2 and cowplot are required for plot.fpca function")
+  }
   # some NULL variable definitions to appease CRAN package checks regarding the use of ggplot2
   fpc_value = type = value = NULL
   
@@ -78,7 +83,7 @@ plot.fpca = function(x, plot_npc = x$npc, var_factor = 2,
   plotDat_list = lapply(1:plot_npc, function(i) {
     # dataset with mean +/- <var_factor>*FPC
     plot_dat = fpc_dat %>%
-      filter(id == i) %>%
+      dplyr::filter(id == i) %>%
       arrange(t) %>%
       dplyr::left_join(mean_dat, by = "t") %>%
       mutate(mean_plusFPC  = response_function(mean + var_factor*fpc_value),
@@ -95,13 +100,17 @@ plot.fpca = function(x, plot_npc = x$npc, var_factor = 2,
                           stringsAsFactors = FALSE) %>%
       mutate(type = factor(type, levels = c("mean curve",
                                             paste0("mean + ",var_factor,"*FPC"),
-                                              paste0("mean - ",var_factor,"*FPC"))))
+                                            paste0("mean - ",var_factor,"*FPC"))))
   })
   
   if (is.null(xlim))
-    xlim = dplyr::bind_rows(plotDat_list) %>% pull(t) %>% range()
+    xlim = dplyr::bind_rows(plotDat_list) %>% 
+    dplyr::pull(t) %>% 
+    range()
   if (is.null(ylim))
-    ylim = dplyr::bind_rows(plotDat_list) %>% pull(value) %>% range()
+    ylim = dplyr::bind_rows(plotDat_list) %>% 
+    dplyr::pull(value) %>% 
+    range()
   
   # create list with one plot per FPC
   plot_list = lapply(1:plot_npc, function(i) {
@@ -115,19 +124,23 @@ plot.fpca = function(x, plot_npc = x$npc, var_factor = 2,
     symbolMinus_dat = symbol_dat %>% filter(type == paste0("mean - ",var_factor,"*FPC"))
     
     # plot
-    ggplot(mapping = aes(x = t, y = value, col = type, lty = type)) +
-      geom_line(data = plotDat_list[[i]]) +
-      scale_color_manual(values = c("black", "gray70", "gray60")) +
-      scale_linetype_manual(values = c(1, 2, 3)) +
-      geom_text(data = symbolPlus_dat,  label = "+", size = 7, col = "gray40") +
-      geom_text(data = symbolMinus_dat, label = "-", size = 7, col = "gray40") +
[…]
