+++ NAMESPACE
@@ -14,12 +14,10 @@ export(registr)
 export(simulate_functional_data)
 export(simulate_unregistered_curves)
 import(dplyr)
-import(ggplot2)
 import(lme4)
 import(mgcv)
 importFrom(MASS,ginv)
 importFrom(Rcpp,evalCpp)
-importFrom(cowplot,plot_grid)
 importFrom(dplyr,arrange)
 importFrom(dplyr,bind_rows)
 importFrom(dplyr,filter)

[…]
