+++ R/registr.R
@@ -62,6 +62,7 @@
 #' call is parallelized by using \code{parallel::mclapply} (for Unix-based
 #' systems) or \code{parallel::parLapply} (for Windows). Defaults to 1,
 #' no parallelized call.
+#' @param verbose print diagnostic messages.
 #' @param ... additional arguments passed to or from other functions
 #' 
 #' @return An list containing:
@@ -100,6 +101,8 @@
 #'                           family = "gaussian", gradient = TRUE,
 #'                           incompleteness = NULL)
 #' if (requireNamespace("ggplot2", quietly = TRUE)) {
+#'   library(ggplot2)
+#' 
 #'   ggplot(register_step2a$Y, aes(x = tstar, y = index, group = id)) +
 #'     geom_line(alpha = 0.2) +
 #'     ggtitle("Estimated warping functions")
@@ -159,7 +162,9 @@ registr = function(obj = NULL, Y = NULL, Kt = 8, Kh = 4, family = "gaussian", gr
                    Y_template = NULL,
                    beta = NULL, t_min = NULL, t_max = NULL, row_obj = NULL,
                    periodic = FALSE, warping = "nonparametric",
-                   gamma_scales = NULL, cores = 1L, ...){
+                   gamma_scales = NULL, cores = 1L, 
+                   verbose = TRUE, 
+                   ...){
   
   if (!is.null(incompleteness)) {
     if (warping != "nonparametric") {
@@ -287,15 +292,18 @@ registr = function(obj = NULL, Y = NULL, Kt = 8, Kh = 4, family = "gaussian", gr
   
   # main function call
   if (cores == 1) { # serial call
-    if (requireNamespace("pbapply", quietly = TRUE)) {
-      results_list = pbapply::pblapply(1:I, registr_oneCurve, arg_list, ...)
+    if (requireNamespace("pbapply", quietly = TRUE) && verbose) {
+      results_list = pbapply::pblapply(1:I, registr_oneCurve, arg_list, ...,
+                                       verbose = verbose > 1)
     } else {
-      results_list = lapply(1:I, registr_oneCurve, arg_list, ...)
+      results_list = lapply(1:I, registr_oneCurve, arg_list, ...,
+                            verbose = verbose > 1)
     }
     
   } else if (.Platform$OS.type == "unix") { # parallelized call on Unix-based systems
     results_list = parallel::mclapply(1:I, registr_oneCurve, arg_list, ...,
-                                      mc.cores = cores)
+                                      mc.cores = cores,
+                                      verbose = verbose > 1)
     
   } else { # parallelized call on Windows
     local_cluster = parallel::makePSOCKcluster(rep("localhost", cores)) # set up cluster
@@ -304,7 +312,8 @@ registr = function(obj = NULL, Y = NULL, Kt = 8, Kh = 4, family = "gaussian", gr
     parallel::clusterEvalQ(cl = local_cluster, c(library(registr), library(stats)))
     
     results_list = parallel::parLapply(cl  = local_cluster, X = 1:I,
-                                       fun = registr_oneCurve, arg_list, ...)
+                                       fun = registr_oneCurve, arg_list, ...,
+                                       verbose = verbose > 1)
     
     parallel::stopCluster(cl = local_cluster) # close cluster
   }
@@ -344,6 +353,7 @@ registr = function(obj = NULL, Y = NULL, Kt = 8, Kh = 4, family = "gaussian", gr
 #' @param i Numeric index of the curve under focus.
[…]
