### 2021-02-03 f271dae
added start of new onecurve function
