+++ R/registr.R
@@ -385,21 +385,40 @@ registr = function(obj = NULL, Y = NULL, Kt = 8, Kh = 4, family = "gaussian", gr
 #' @importFrom splines bs
 #' @importFrom stats constrOptim Gamma
 #' 
-registr_oneCurve = function(i, arg_list, ..., verbose = TRUE) {
+registr_oneCurve = function(
+  obj = NULL,
+  Y = NULL, 
+  Kt = 8, 
+  Kh = 4, 
+  family = "gaussian",
+  gradient = TRUE,
+  incompleteness = NULL, 
+  lambda_inc = NULL,
+  beta = NULL, 
+  t_min = NULL, 
+  t_max = NULL,
+  periodic = FALSE, 
+  warping = "nonparametric",
+  gamma_scales = NULL,
+  global_knots = NULL,
+  mean_coefs = NULL,
+  ...,
+  verbose = TRUE) {
   
-  t_min_i       = arg_list$Y$tstar[arg_list$rows$first_row[i]]
-  t_max_i       = arg_list$Y$tstar[arg_list$rows$last_row[i]]
-  Y_cropped     = arg_list$Y[arg_list$Y$tstar >= t_min_i & arg_list$Y$tstar <= t_max_i,]
-  tstar_cropped = Y_cropped$tstar
-  rows_i        = which(Y_cropped$id == arg_list$rows$id[i])
-  Y_i           = Y_cropped$value[rows_i]
+  t_range_i     = range(Y$tstar)
+  t_min_i       = t_range_i[1]
+  t_max_i       = t_range_i[2]
+  Y_i           = Y
+  #! needs to change
+  tstar_cropped = Y$tstar
+  rm(Y)
   D_i           = length(Y_i)
   
   # spline basis on the curve-specific time interval.
   # Pre-calculate the knots (for setting up the curve-specific Theta_h spline
   # bases in registr_oneCurve) similarly to splines::bs to make the final splines::bs call faster.
   degree          = 3 # cubic splines
-  n_innerKnots    = arg_list$Kh - (1 + degree)
+  n_innerKnots    = Kh - (1 + degree)
   if (n_innerKnots <= 0) {
     hinv_innerKnots_i = NULL
   } else if (n_innerKnots > 0) {
@@ -409,32 +428,32 @@ registr_oneCurve = function(i, arg_list, ..., verbose = TRUE) {
   if (verbose) {
     message("Getting Spline Basis")
   }
-  tstar_bs_i      = c(tstar_cropped[rows_i], range(tstar_cropped))
-  Theta_h_cropped = splines::bs(tstar_bs_i, knots = hinv_innerKnots_i, intercept = TRUE)
-  Theta_h_i       = Theta_h_cropped[1:length(rows_i),]
+  tstar_bs_i      = c(Y$tstar, range(tstar_cropped))
+  Theta_h_i       = splines::bs(tstar_bs_i, knots = hinv_innerKnots_i, intercept = TRUE)
+  Theta_h_i       = Theta_h_i[1:nrow(Y),]
   
   # start parameters
-  if (is.null(arg_list$beta)) { # newly initialize start parameters
-    t_vec_i     = Y_cropped$tstar[rows_i]
-    beta_full_i = initial_params(warping = arg_list$warping,
+  if (is.null(beta)) { # newly initialize start parameters
+    t_vec_i     = Y$tstar
+    beta_full_i = initial_params(warping = warping,
                                  K       = Theta_h_i,
                                  t_vec   = t_vec_i)
     
   } else { # use the last estimates (from the joint approach) as start parameters
-    beta_full_i = arg_list$beta[, i]
+    beta_full_i = beta
   }
   # account the beta vector for potential incompleteness constraints
[…]
