### 2021-02-03 3a35ba0
register curve is now differnt implementation for splitting up Y
