+++ R/registr.R
@@ -288,35 +288,36 @@ registr = function(obj = NULL, Y = NULL, Kt = 8, Kh = 4, family = "gaussian", gr
                   t_max          = t_max,          rows         = rows,
                   periodic       = periodic,       warping      = warping,
                   global_knots   = global_knots,   mean_coefs   = mean_coefs,
-                  gamma_scales   = gamma_scales)
+                  gamma_scales   = gamma_scales,
+                  ...)
   
   # main function call
+  ids = Y$id
+  arg_list$Y = NULL
+  arg_list$rows = NULL
+  Y = split(Y, ids)
+  args = arg_list
   if (cores == 1) { # serial call
-    if (FALSE) {
-      ids = Y$id
-      arg_list$Y = NULL
-      arg_list$rows = NULL
-      Y = split(Y, ids)
+    if (requireNamespace("pbapply", quietly = TRUE) && verbose) {
       results_list = pbapply::pblapply(Y, function(r) {
-        args = arg_list
         args$Y = r
         args$verbose = verbose > 1
         do.call(registr_oneCurve, args = args)
       })
-      Y = unsplit(Y, ids)
-    }
-    if (requireNamespace("pbapply", quietly = TRUE) && verbose) {
-      results_list = pbapply::pblapply(1:I, registr_oneCurve, arg_list, ...,
-                                       verbose = verbose > 1)
     } else {
-      results_list = lapply(1:I, registr_oneCurve, arg_list, ...,
-                            verbose = verbose > 1)
+      results_list = lapply(Y, function(r) {
+        args$Y = r
+        args$verbose = verbose > 1
+        do.call(registr_oneCurve, args = args)
+      })
     }
     
   } else if (.Platform$OS.type == "unix") { # parallelized call on Unix-based systems
-    results_list = parallel::mclapply(1:I, registr_oneCurve, arg_list, ...,
-                                      mc.cores = cores,
-                                      verbose = verbose > 1)
+    results_list = parallel::mclapply(Y, function(r) {
+      args$Y = r
+      args$verbose = verbose > 1
+      do.call(registr_oneCurve, args = args)
+    }, mc.cores = cores)
     
   } else { # parallelized call on Windows
     local_cluster = parallel::makePSOCKcluster(rep("localhost", cores)) # set up cluster
@@ -324,12 +325,18 @@ registr = function(obj = NULL, Y = NULL, Kt = 8, Kh = 4, family = "gaussian", gr
     parallel::clusterExport(cl = local_cluster, c("arg_list"), envir=environment())
     parallel::clusterEvalQ(cl = local_cluster, c(library(registr), library(stats)))
     
-    results_list = parallel::parLapply(cl  = local_cluster, X = 1:I,
-                                       fun = registr_oneCurve, arg_list, ...,
-                                       verbose = verbose > 1)
+    results_list = parallel::parLapply(
+      cl  = local_cluster,
+      X = Y, fun = function(r) {
+      args$Y = r
+      args$verbose = verbose > 1
+      do.call(registr_oneCurve, args = args)
+    })    
     
     parallel::stopCluster(cl = local_cluster) # close cluster
   }
+  Y = dplyr::bind_rows(Y)
+  rm(ids)
   
   # gather the results
   hinv_innerKnots        = lapply(results_list, function(x) x$hinv_innerKnots)
@@ -363,11 +370,7 @@ registr = function(obj = NULL, Y = NULL, Kt = 8, Kh = 4, family = "gaussian", gr
 #' It performs the main optimization step with \code{constrOptim} for the
 #' registration of one curve.
 #' 
-#' @param i Numeric index of the curve under focus.
[…]
