### 2021-02-03 b4194bf
fixes for vignettes and ggplot2 and fixing oneCurve because more subject-specific measures need to be passed
