+++ R/registr.R
@@ -246,6 +246,9 @@ registr = function(obj = NULL, Y = NULL, Kt = 8, Kh = 4, family = "gaussian", gr
   } else { # template function = mean of all curves
     
     if (!is.null(Y_template)) {
+      if (verbose) {
+        message("Registr: Extracting Y_template")
+      }
       if (!all(c("id", "index", "value") %in% names(Y_template))) {
         stop("Y_template must have variables 'id', 'index', and 'value'.")
       } else if (!identical(range(Y_template$index), range(Y$index))) {
@@ -257,6 +260,9 @@ registr = function(obj = NULL, Y = NULL, Kt = 8, Kh = 4, family = "gaussian", gr
       mean_dat = Y
     }
     
+    if (verbose) {
+      message("Registr: Getting Knots and basis functions")
+    }
     if (periodic) {
       # if periodic, then we want more global knots, because the resulting object from pbs 
       # only has (knots+intercept) columns.
@@ -296,44 +302,70 @@ registr = function(obj = NULL, Y = NULL, Kt = 8, Kh = 4, family = "gaussian", gr
   arg_list$Y = NULL
   arg_list$rows = NULL
   Y = split(Y, ids)
+  if (!is.null(beta)) {
+    beta_list = apply(beta, 2, list)
+    beta_list = lapply(beta_list, unlist)
+    Y = mapply(function(y, b) {
+      attr(y, "beta") = b
+      y
+    }, Y, beta_list, SIMPLIFY = FALSE)
+    rm(beta_list)
+  }
+  if (!is.null(mean_coefs)) {
+    if (is.matrix(mean_coefs)) {
+      mc_list = apply(mean_coefs, 2, list)
+      mc_list = lapply(mc_list, unlist)
+    } else {
+      mc_list = lapply(1:length(Y), function(blah) {
+        mean_coefs
+      })
+    }
+    Y = mapply(function(y, b) {
+      attr(y, "mean_coefs") = b
+      y
+    }, Y, mc_list, SIMPLIFY = FALSE)
+    rm(mc_list)
+  }  
   args = arg_list
+  rm(arg_list)
+  if (verbose) {
+    message("Registr: Running individual curves")
+  }
+  run_one_curve = function(r) {
+    args$Y = r
+    args$beta = attr(r, "beta")
+    args$mean_coefs = attr(r, "mean_coefs")
+    args$verbose = verbose > 1
+    do.call(registr_oneCurve, args = args)
+  }
   if (cores == 1) { # serial call
     if (requireNamespace("pbapply", quietly = TRUE) && verbose) {
-      results_list = pbapply::pblapply(Y, function(r) {
-        args$Y = r
-        args$verbose = verbose > 1
-        do.call(registr_oneCurve, args = args)
-      })
+      results_list = pbapply::pblapply(Y, run_one_curve)
     } else {
-      results_list = lapply(Y, function(r) {
-        args$Y = r
-        args$verbose = verbose > 1
-        do.call(registr_oneCurve, args = args)
-      })
+      results_list = lapply(Y, run_one_curve)
     }
     
   } else if (.Platform$OS.type == "unix") { # parallelized call on Unix-based systems
-    results_list = parallel::mclapply(Y, function(r) {
-      args$Y = r
-      args$verbose = verbose > 1
-      do.call(registr_oneCurve, args = args)
-    }, mc.cores = cores)
+    results_list = parallel::mclapply(Y, run_one_curve, mc.cores = cores)
     
   } else { # parallelized call on Windows
     local_cluster = parallel::makePSOCKcluster(rep("localhost", cores)) # set up cluster
     # export functions and packages to the cluster
[…]
