+++ vignettes/registr.Rmd
@@ -50,9 +50,12 @@ found in the separate vignette `"incomplete_curves"`.
 
 ```{r load_libraries, echo = FALSE}
 library(registr)
-library(ggplot2)
 library(dplyr)
-theme_set(theme_bw())
+have_ggplot2 = requireNamespace("ggplot2", quietly = TRUE)
+if (have_ggplot2)  {
+  library(ggplot2)
+  theme_set(theme_bw())
+}
 ```
 
 
@@ -118,7 +121,7 @@ The resulting object,`registration_data`, is a data frame with variables `id`, `
 
 The variables `index` and `t` are both time grids. Evaluated on the grid `index` the data is unregistered, and on the grid `t` the data is registered. Registered and unregistered curves are plotted below.
 
-```{r plot_sim2, echo = FALSE, fig.show='hold'}
+```{r plot_sim2, echo = FALSE, fig.show='hold', eval = have_ggplot2}
 
 registration_data %>%
 	ggplot(aes(index, plogis(latent_mean), group = id)) + theme_bw() + 
@@ -155,22 +158,25 @@ pc_df = data.frame(pop_mean = fpca_data$alpha,
 									 index = seq(0, 1, length.out = 200),
 									 id = 1)
 
-ggplot(Y, aes(index, latent_mean, group = id)) + theme_bw() +
-	geom_line(alpha = 0.25) + geom_line(data = pc_df, aes(y = pop_mean), color = "red") 
-
-ggplot(pc_df, aes(index, psi1)) + theme_bw() + geom_line(color = "blue") 
-ggplot(pc_df, aes(index, psi2)) + theme_bw() + geom_line(color = "blue") 
-
+if (have_ggplot2) {
+  ggplot(Y, aes(index, latent_mean, group = id)) + theme_bw() +
+    geom_line(alpha = 0.25) + geom_line(data = pc_df, aes(y = pop_mean), color = "red") 
+  
+  ggplot(pc_df, aes(index, psi1)) + theme_bw() + geom_line(color = "blue") 
+  ggplot(pc_df, aes(index, psi2)) + theme_bw() + geom_line(color = "blue") 
+}
 ```
 
 The left panel of the figure above shows the latent means for each subject, along with the population-level mean,  $\alpha(t)$, in red. The middle and right panels show the first and second principal components, $\psi_1(t)$ and $\psi_2(t)$, respectively. Using the $logit^{-1}(\cdot)$ function we can convert the subject-specific means to probabilities; these probabilities are used to generate the binary values. Binary values and latent probability curve for one subject in the dataset is shown below.
 
 ```{r plot2_sim1, echo = FALSE}
-Y %>%
-	filter(id == 7) %>%
-	ggplot(aes(index, value)) + theme_bw() +
-	geom_point(alpha = 0.75, size = 0.25) + geom_line(aes(y = plogis(latent_mean))) +
-	labs(y = "Pr(Y = 1)")
+if (have_ggplot2) {
+  Y %>%
+    filter(id == 7) %>%
+    ggplot(aes(index, value)) + theme_bw() +
+    geom_point(alpha = 0.75, size = 0.25) + geom_line(aes(y = plogis(latent_mean))) +
+    labs(y = "Pr(Y = 1)")
+}
 ```
 
 We can alter the score variance for the principal components using the arguments `lambda1` and `lambda2`. The default setting is for all subjects to have the same number of time points. However, by specifying `vary_D = TRUE`, we can generate data with uneven grid lengths for each subject.
@@ -185,35 +191,36 @@ The type of GFPCA is specified by the argument `fpca_type`, either calling the v
[…]

[…]
