### 2021-02-03 94ca1f1
added the necessary arguments for makevars and added to the docs
