### 2021-02-06 72b7234
added more verbosity
