+++ R/fpca_gauss.R
@@ -18,6 +18,7 @@
 #' @param periodic If TRUE, uses periodic b-spline basis functions. Default is FALSE.
 #' @param error_thresh Error threshold to end iterations. Defaults to 0.0001.
 #' @param ... Additional arguments passed to or from other functions
+#' @param verbose print diagnostic messages.
 #' 
 #' @author Julia Wrobel \email{julia.wrobel@@cuanschutz.edu},
 #' Jeff Goldsmith \email{ajg2202@@cumc.columbia.edu}
@@ -54,7 +55,8 @@
 #' 
 fpca_gauss = function(Y, npc = 1, Kt = 8, maxiter = 20, t_min = NULL, t_max = NULL, 
 											print.iter = FALSE, row_obj= NULL, seed = 1988, periodic = FALSE, 
-											error_thresh = 0.0001, ...){
+											error_thresh = 0.0001, verbose = TRUE, 
+											...){
 	
 	curr_iter = 1
 	error     = rep(NA, maxiter)
