+++ man/fpca_gauss.Rd
@@ -16,6 +16,7 @@ fpca_gauss(
   seed = 1988,
   periodic = FALSE,
   error_thresh = 1e-04,
+  verbose = TRUE,
   ...
 )
 }
@@ -42,6 +43,8 @@ Keep this NULL if you are using standalone \code{register} function.}
 \item{periodic}{If TRUE, uses periodic b-spline basis functions. Default is FALSE.}
 
 \item{error_thresh}{Error threshold to end iterations. Defaults to 0.0001.}
+
+\item{verbose}{print diagnostic messages.}
 
 \item{...}{Additional arguments passed to or from other functions}
 }
