### 2021-02-06 ce9cbbc
allowing subsampling for mean coefficient vector
