### 2021-02-08 b10843d
implemented subsampling for bfpca, changed default method for fastglm
