### 2021-02-08 44c64dd
added some checks for sampling ids
