### 2021-02-08 4efc784
made print.iter the default for bfpca
