### 2021-02-08 #2964 @mbobra
Thanks for the ping, @bauer-alex! I reached out to both reviewers via e-mail. We at JOSS do give reviewers six weeks to complete their review (JOSS is operating on a reduced service mode), so the reviewers still have some time.
