### 2021-02-24 3f2cf5a
had to remove rows from the data
