### 2021-03-03 e09f12d
new adaptive penalty for incomplete curves
