### 2021-03-08 9d298d3
improve cov_hall
- center first, then compute crossprods and smooth
- use bam for large data
- don't set up humongous crossprod matrix for irregular data
- posdefify returned covariance estimate
