+++ NAMESPACE
@@ -18,6 +18,7 @@ import(ggplot2)
 import(lme4)
 import(mgcv)
 importFrom(MASS,ginv)
+importFrom(Matrix,nearPD)
 importFrom(Rcpp,evalCpp)
 importFrom(cowplot,plot_grid)
 importFrom(dplyr,arrange)
@@ -25,6 +26,7 @@ importFrom(dplyr,bind_rows)
 importFrom(dplyr,filter)
 importFrom(dplyr,group_by)
 importFrom(dplyr,mutate)
+importFrom(dplyr,n_distinct)
 importFrom(dplyr,select)
 importFrom(dplyr,ungroup)
 importFrom(gamm4,gamm4)
@@ -36,6 +38,8 @@ importFrom(parallel,mclapply)
 importFrom(parallel,parLapply)
 importFrom(parallel,stopCluster)
 importFrom(pbs,pbs)
+importFrom(purrr,cross_df)
+importFrom(purrr,map)
 importFrom(splines,bs)
 importFrom(splines,splineDesign)
 importFrom(stats,Gamma)
@@ -45,13 +49,17 @@ importFrom(stats,constrOptim)
 importFrom(stats,dnorm)
 importFrom(stats,formula)
 importFrom(stats,glm)
+importFrom(stats,na.omit)
 importFrom(stats,pbeta)
 importFrom(stats,plogis)
+importFrom(stats,predict)
 importFrom(stats,quantile)
 importFrom(stats,rbinom)
 importFrom(stats,rnorm)
 importFrom(stats,runif)
 importFrom(tidyr,gather)
+importFrom(tidyr,nest)
 importFrom(tidyr,spread)
+importFrom(tidyr,unnest)
 importFrom(utils,tail)
 useDynLib(registr)
