### 2021-03-15 4468720
Revert "Merge pull request #35 from muschellij2/master"
This reverts commit 0a7be3fef680705c43dd36940e7be3711ccbc21b, reversing
changes made to 7ad687565ee0e1972c9938de52b29f724d84bd0f.

Commit broke build on Linux hosts.
