+++ man/plot.fpca.Rd
@@ -57,11 +57,7 @@ only the case for FPCA objects created with \code{\link{gfpca_twoStep}}.
 data(growth_incomplete)
 
 fpca_obj = fpca_gauss(Y = growth_incomplete, npc = 2)
-if (requireNamespace("ggplot2", quietly = TRUE) &&
-requireNamespace("cowplot", quietly = TRUE)) {
-library(ggplot2)
 plot(fpca_obj)
-}
 
 }
 \author{

[…]
