### 2021-03-17 6226e20
Include the merge request again by reverting "Revert "Merge pull request #35 from muschellij2/master""
This reverts commit 4468720cfae9122f7016b613ce4781a4e395519d.
