+++ R/registr.R
@@ -1,3 +1,202 @@
+registr_helper = function(obj = NULL, Y = NULL, Kt = 8, Kh = 4, family = "gaussian", gradient = TRUE,
+                          incompleteness = NULL, lambda_inc = NULL,
+                          Y_template = NULL,
+                          beta = NULL, t_min = NULL, t_max = NULL, row_obj = NULL,
+                          periodic = FALSE, warping = "nonparametric",
+                          gamma_scales = NULL, cores = 1L,  subsample = TRUE,
+                          verbose = TRUE,
+                          ...){ 
+  if (!is.null(incompleteness)) {
+    if (warping != "nonparametric") {
+      stop("The functionality for incomplete curves is only available for 'warping = 'nonparametric''")
+    }
+    if (!(incompleteness %in% c("leading","trailing","full"))) {
+      stop("'incompleteness' must be either 'leading', 'trailing' or 'full'.")
+    }
+    if ((is.null(lambda_inc) || (lambda_inc < 0))) {
+      stop("For incomplete curves the penalization parameter 'lambda_inc' has to be set to some nonnegative value.")
+    }
+  }
+  
+  if (is.null(Y)) { 
+    Y = obj$Y
+  }
+  
+  if (family == "gamma" & any(Y$value <= 0)) {
+    stop("family = 'gamma' can only be applied to strictly positive data.")
+  } else if (family == "poisson" & any(Y$value < 0)) {
+    stop("family = 'poisson' can only be applied to nonnegative data.")
+  }
+  
+  if(is.null(obj)) { 
+    if(warping == "nonparametric"){
+      Y$tstar = Y$index
+      
+    } else if(warping == "piecewise_linear2"){
+      # scale time to 0, 1 for parametric warping
+      if(is.null(Y$index_scaled)){
+        stop("For warping = piecewise_linear2, need an index_scaled variable that ranges from 0 to 1.")
+      }
+      Y$tstar = Y$index_scaled
+    }
+  }
+  
+  if (is.null(row_obj)) {
+    data = data_clean(Y)
+    Y    = data$Y
+    rows = data$Y_rows
+    I    = data$I
+  } else{
+    rows = row_obj
+    I    = nrow(rows)
+  }
+  
+  if (Kh < 4) {
+    stop("Kh must be greater than or equal to 4.")
+  }
+  if (Kt < 4) {
+    stop("Kt must be greater than or equal to 4.")
+  }
+  
+  if(!(warping %in% c("nonparametric", "piecewise_linear2"))){
+    stop("warping argument can only take values of nonparametric or piecewise_linear2")
+  }
+  
+  if (gradient & !(family %in% c("gaussian","binomial"))) {
+    warning("gradient = TRUE is only available for families 'gaussian' and 'binomial'. Setting gradient = FALSE.")
+    gradient = FALSE
+  } else if (gradient & periodic){
+    warning("gradient = TRUE is only available for periodic = FALSE. Setting gradient = FALSE.")
+    gradient = FALSE
+  } else if (gradient & warping != "nonparametric"){
+    warning("gradient = TRUE is only available for warping = nonparametric. Setting gradient = FALSE.")
+    gradient = FALSE
+  }
+  
+  if (is.null(t_min)) { t_min = min(Y$tstar) }
+  if (is.null(t_max)) { t_max = max(Y$tstar) }
+  stopifnot(!is.na(t_min),
+            !is.na(t_max))
+  
+  if (!is.null(obj)) { # template function = GFPCA representation
+    global_knots = obj$knots
+    mean_coefs   = obj$subject_coefs
+    
+  } else { # template function = mean of all curves
+    
[…]
