### 2021-03-17 ded49b6
made some unit tests a bit faster
