+++ tests/testthat/test-register_fpca.R
@@ -8,9 +8,11 @@ test_that("code only accepts supported family of distributions", {
 	register_fpca(Y, family = "gaussian", max_iterations = 5)
 								 
 	Y$value = Y$value + 1 # make y values strictly positive for gamma and poisson family
-	expect_warning(register_fpca(Y, family = "gamma", max_iterations = 2, fpca_index_significantDigits = 2),
+	expect_warning(register_fpca(Y, family = "gamma", fpca_type = "two-step", gradient = FALSE,
+	                             max_iterations = 1, fpca_index_significantDigits = 2),
 								 "Convergence not reached. Try increasing max_iterations.")
-	expect_warning(register_fpca(Y, family = "poisson", max_iterations = 2, fpca_index_significantDigits = 2),
+	expect_warning(register_fpca(Y, family = "poisson", fpca_type = "two-step", gradient = FALSE,
+	                             max_iterations = 1, fpca_index_significantDigits = 2),
 								 "Convergence not reached. Try increasing max_iterations.")
 	
 	expect_error(register_fpca(Y, family = "exponential"),
@@ -22,7 +24,9 @@ test_that("code only accepts supported family of distributions", {
 test_that("registering binary data throws no errors",{
 	Y = simulate_unregistered_curves(seed = 10001)
 	register_fpca(Y, family = "binomial", fpca_type = "variationalEM", max_iterations = 2)
-	register_fpca(Y, family = "binomial", fpca_type = "two-step", max_iterations = 2)
+	expect_warning(register_fpca(Y, family = "binomial", fpca_type = "two-step",
+	                             max_iterations = 1, fpca_index_significantDigits = 2),
+	               "Convergence not reached. Try increasing max_iterations.")
 	
 	dat = Y$index
 	expect_error(register_fpca(dat, family = "binomial"), 
@@ -53,7 +57,9 @@ test_that("registering poisson data throws no errors",{
 test_that("register_fpca output is a list with non-null items and class registration",{
 	Y = simulate_unregistered_curves(seed = 1001)
 	registr_object1 = register_fpca(Y, family = "binomial", fpca_type = "variationalEM", max_iterations = 2)
-	registr_object2 = register_fpca(Y, family = "binomial", fpca_type = "two-step", max_iterations = 2)
+	expect_warning({
+	  registr_object2 = register_fpca(Y, family = "binomial", fpca_type = "two-step", max_iterations = 1, fpca_index_significantDigits = 2)
+	}, "Convergence not reached. Try increasing max_iterations.")
 	
 	expect_equal(class(registr_object1), "registration")
 	expect_equal(class(registr_object2), "registration")
