### 2021-03-17 9f3cef9
minor code cleanup
