### 2021-03-17 a8018f1
some alterations to plot.fpca
