### 2021-03-17 d3bc392
added adaptive part to the incomplete curve penalty to account for different numbers of measurements
