+++ paper/paper.md
@@ -104,7 +104,7 @@ for curve $i$. For a setting with full incompleteness this results in
 
 $$
 \begin{aligned}
-\ell_{\text{pen}}(i) &= \ell(i) - \lambda \cdot \text{pen}(i), \\
+\ell_{\text{pen}}(i) &= \ell(i) - \lambda \cdot n_i \cdot \text{pen}(i), \\
 \text{with} \ \ \ 
 \text{pen}(i) &= \left( [\hat{h}_i^{-1}(t_{max,i}^*) - \hat{h}_i^{-1}(t_{min,i}^*)] - [t_{max,i}^* - t_{min,i}^*] \right)^2,
 \end{aligned}
@@ -117,6 +117,9 @@ For leading incompleteness with $h_i^{-1}(t_{max,i}^*) = t_{max,i}^* \forall i$ 
 $\text{pen}(i) = \left(\hat{h}_i^{-1}(t_{min,i}^*) - t_{min,i}^*\right)^2$, and for trailing incompleteness with
 $h_i^{-1}(t_{min,i}^*) = t_{min,i}^* \forall i$ to
 $\text{pen}(i) = \left(\hat{h}_i^{-1}(t_{max,i}^*) - t_{max,i}^*\right)^2$.
+The penalization term is scaled by the number of measurements $n_i$ of curve $i$
+to ensure a similar impact of the penalization for curves with different numbers
+of measurements.
 In practical settings, $\lambda$ has to be set manually to specify which kinds of
 warpings are deemed unrealistic and should be prevented.
 The choice of $\lambda$ should be based on subject knowledge by comparing
