+++ man/crossprods_regular.Rd
@@ -1,0 +1,17 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/gfpca_covHall.R
+\name{crossprods_regular}
+\alias{crossprods_regular}
+\title{Crossproduct computation for mostly regular grids}
+\usage{
+crossprods_regular(Y)
+}
+\arguments{
+\item{Y}{Dataframe with the centered observations.
+Should have values id, centered, index.}
+}
+\description{
+Compute the crossproduct in a fast way for mostly regular grids
+(index values are mostly *not* unique).
+Only used internally in \code{cov_hall()}.
+}
