### 2021-03-17 faf7d39
unit tests for crossproduct calculation + made other unit tests a bit faster
