+++ tests/testthat/test-register_fpca.R
@@ -43,14 +43,16 @@ test_that("registering gaussian data throws no errors",{
 test_that("registering gamma data throws no errors",{
 	Y       = simulate_unregistered_curves(seed = 10001)
 	Y$value = Y$value + 1 # make y values strictly positive for gamma family
-	expect_warning(register_fpca(Y, family = "gamma", max_iterations = 2, fpca_index_significantDigits = 2),
+	expect_warning(register_fpca(Y, family = "gamma", fpca_type = "two-step", gradient = FALSE,
+	                             max_iterations = 1, fpca_index_significantDigits = 2),
 								 "Convergence not reached. Try increasing max_iterations.")
 })
 
 test_that("registering poisson data throws no errors",{
 	Y       = simulate_unregistered_curves(seed = 10001)
 	Y$value = Y$value + 1 # make y values nonnegative for poisson family
-	expect_warning(register_fpca(Y, family = "poisson", max_iterations = 2, fpca_index_significantDigits = 2),
+	expect_warning(register_fpca(Y, family = "poisson", fpca_type = "two-step", gradient = FALSE,
+	                             max_iterations = 1, fpca_index_significantDigits = 2),
 								 "Convergence not reached. Try increasing max_iterations.")
 })
 
@@ -77,15 +79,15 @@ test_that("register_fpca function forces gradient = FALSE in some situations",{
 	data    = data_clean(Y)
 	Y       = data$Y
 	
-	expect_warning(register_fpca(Y, family = "binomial", warping = "piecewise_linear2", gradient = TRUE, max_iterations = 2), 
+	expect_warning(register_fpca(Y, family = "binomial", warping = "piecewise_linear2", gradient = TRUE, max_iterations = 1), 
 								 "gradient = TRUE is only available for warping = nonparametric. Setting gradient = FALSE.")
-	expect_warning(register_fpca(Y, family = "binomial", periodic = TRUE, gradient = TRUE, max_iterations = 2), 
+	expect_warning(register_fpca(Y, family = "binomial", periodic = TRUE, gradient = TRUE, max_iterations = 1), 
 								 "gradient = TRUE is only available for periodic = FALSE. Setting gradient = FALSE.")
 	
 	Y$value = Y$value + 1 # make y values strictly positive for gamma and poisson family
-	expect_warning(register_fpca(Y, family = "gamma", max_iterations = 2, fpca_index_significantDigits = 2), 
+	expect_warning(register_fpca(Y, family = "gamma", fpca_type = "two-step", gradient = TRUE, max_iterations = 1, fpca_index_significantDigits = 2), 
 								 "gradient = TRUE is only available for families 'gaussian' and 'binomial'. Setting gradient = FALSE.")
-	expect_warning(register_fpca(Y, family = "poisson", max_iterations = 2, fpca_index_significantDigits = 2), 
+	expect_warning(register_fpca(Y, family = "poisson", fpca_type = "two-step", gradient = TRUE, max_iterations = 1, fpca_index_significantDigits = 2), 
 								 "gradient = TRUE is only available for families 'gaussian' and 'binomial'. Setting gradient = FALSE.")
 })
 
@@ -93,9 +95,11 @@ test_that("register_fpca function forces fpca_type = 'two-step' for families oth
 	Y = registr::growth_incomplete
 	Y$value = Y$value + 1 # make y values strictly positive for gamma and poisson family
 	
-	expect_warning(register_fpca(Y, family = "gamma", fpca_type = "variationalEM", max_iterations = 2, fpca_index_significantDigits = 2),
+	expect_warning(register_fpca(Y, family = "gamma", fpca_type = "variationalEM", gradient = FALSE,
[…]
