### 2021-03-17 4635c92
minor fixes of the verbose functionality + set default of print.iter to FALSE in bfpca, similar to fpca_gauss
