+++ man/bfpca.Rd
@@ -11,7 +11,7 @@ bfpca(
   maxiter = 50,
   t_min = NULL,
   t_max = NULL,
-  print.iter = TRUE,
+  print.iter = FALSE,
   row_obj = NULL,
   seed = 1988,
   periodic = FALSE,
