### 2021-03-17 ddfa3b0
minor fixes in readme
