+++ README.Rmd
@@ -10,7 +10,7 @@ output: github_document
 [![](https://www.r-pkg.org/badges/version/registr)](https://cran.r-project.org/package=registr)
 [![](http://cranlogs.r-pkg.org/badges/grand-total/registr?color=green)](https://cran.r-project.org/package=registr)
 [![](https://travis-ci.org/julia-wrobel/registr.svg?branch=master)](https://travis-ci.org/julia-wrobel/registr)
-[![AppVeyor build status](https://ci.appveyor.com/api/projects/status/github/muschellij2/registr?branch=master&svg=true)](https://ci.appveyor.com/project/muschellij2/registr)
+[![AppVeyor build status](https://ci.appveyor.com/api/projects/status/github/julia-wrobel/registr?branch=master&svg=true)](https://ci.appveyor.com/project/julia-wrobel/registr)
 [![Codecov test coverage](https://codecov.io/gh/julia-wrobel/registr/branch/master/graph/badge.svg)](https://codecov.io/gh/julia-wrobel/registr/coverage.svg?branch=master)
 [![status](http://joss.theoj.org/papers/9c40c4f0ede1827cc5a9430c625d6494/status.svg)](http://joss.theoj.org/papers/9c40c4f0ede1827cc5a9430c625d6494)
  [![R-CMD-check](https://github.com/julia-wrobel/registr/workflows/R-CMD-check/badge.svg)](https://github.com/julia-wrobel/registr/actions)
@@ -135,7 +135,7 @@ Our software registers curves by estimating $t$. For this we use the function `r
 
 ```{r register_data, echo = TRUE, message = TRUE}
 binary_registration = register_fpca(Y = registration_data, family = "binomial", 
-                                    Kt = 6, Kh = 4, npc  = 1)
+                                    Kt = 6, Kh = 4, npc  = 1, verbose = FALSE)
 ```
 
 The plot below shows unregistered, true registered, and estimated registered binary curves for two subjects after fitting our method.
