### 2021-03-18 6d9c7bd
travis.yml: removed a test script since it didn't fix the travis errors
