+++ .travis.yml
@@ -4,6 +4,10 @@ language: R
 sudo: false
 cache: packages
 
+env:
+ global:
+   - _R_CHECK_FORCE_SUGGESTS_: false
+
 r_packages:
   - covr
 
@@ -14,7 +18,3 @@ matrix:
   include:
   - r: devel
   - r: release
-
-# the following was added as of this error:
-# https://stackoverflow.com/questions/62956526/checking-r-code-with-travis-ci-gives-error-the-command-grep-q-r-warning
-script: ! grep -R "WARNING" "${RCHECK_DIR}/00check.log"
