### 2021-03-18 d2fed08
#37 set travis timeout to 30 minutes
