### 2021-03-18 f5231e3
set default of verbose to FALSE + version bump
