+++ man/fpca_gauss.Rd
@@ -17,7 +17,7 @@ fpca_gauss(
   periodic = FALSE,
   error_thresh = 1e-04,
   subsample = TRUE,
-  verbose = TRUE,
+  verbose = FALSE,
   ...
 )
 }
@@ -48,7 +48,7 @@ Keep this NULL if you are using standalone \code{register} function.}
 \item{subsample}{if the number of rows of the data is greater than 
 10 million rows, the `id` values are subsampled to get the mean coefficients.}
 
-\item{verbose}{print diagnostic messages.}
+\item{verbose}{If TRUE, prints diagnostic messages. Defaults to FALSE.}
 
 \item{...}{Additional arguments passed to or from other functions}
 }
