### 2021-03-28 julia-wrobel/registr#38 @qpmnguyen
[Review] JOSS Review Feedback
Hi all,

This is the feedback issue for the current JOSS submission [here](https://github.com/openjournals/joss-reviews/issues/2964).

I think this is a wonderful submission and I would recommend it for acceptance for publication in JOSS. The package was really easy to use and is well documented and supported, where I especially appreciate that vignettes list all the important user-facing functions. Sometimes, packages can have too many functions and it's hard to know where to start. Following are just some minor recommendations I have, mostly relating to documentation.

1. I would appreciate a bit more discussion on the choice of distribution family other than gaussian and binomial. I understand that these are the two common types of outcomes that users might face, but the package does support a larger number of available families that should be featured.
2. There is a section in the vignette that includes a discussion on the differences in computational time between using an analytical versus numerical gradient. I think that the discussion should include some information as to when a user might want to use a numerical gradient versus an analytical one, given improvements in computational time.
3. Relatedly, I think the package would also benefit from having a bit more details on computational time performance. Users might not really know what is the limiting factor for computational time for the package (Sample size? Number of time points?). The authors can include computational time for a typical experiment a user might encounter.
4. Given that the package already has real data bundled, it would be interesting to see in the vignette a short analysis of that data set.
5. While I understand that this package implements a novel statistical approach, it would also be great to list packages that are in the same space. Right now the packages referenced in the text are alternative/related implementation of specific parts of the algorithm, rather than a survey of other existing software in R to analyze functional data.

That said, this was a great submission. Feel free to close this issue at your convenience.
