### 2021-03-28 #2964 @qpmnguyen
Thank you very much everyone for your patience. @mbobra I have gone through the checklist and found the submission to have fulfilled all the listed requirements. I recommend acceptance for publication in JOSS. I have also filed an issue with some minor comments [here](https://github.com/julia-wrobel/registr/issues/38)
