### 2021-03-31 d7e54cb
various vignette adaptations wrt points 1-4 of #38
