### 2021-03-31 77aff78
#38 included alternative R packages in the JOSS paper
