+++ paper/paper.bib
@@ -40,3 +40,27 @@
   publisher={Wiley Online Library},
   doi={10.1111/biom.12963}
 }
+
+@Manual{R_fda,
+  title = {fda: Functional Data Analysis},
+  author = {J. O. Ramsay and Spencer Graves and Giles Hooker},
+  year = {2020},
+  note = {R package version 5.1.5.1},
+  url = {https://CRAN.R-project.org/package=fda},
+}
+
+@Manual{R_fdakma,
+  title = {fdakma: Functional Data Analysis: K-Mean Alignment},
+  author = {Alice Parodi and Mirco Patriarca and Laura Sangalli and Piercesare Secchi and Simone Vantini and Valeria Vitelli},
+  year = {2015},
+  note = {R package version 1.2.1},
+  url = {https://CRAN.R-project.org/package=fdakma},
+}
+
+@Manual{R_fdasrvf,
+  title = {fdasrvf: Elastic Functional Data Analysis},
+  author = {J. Derek Tucker},
+  year = {2020},
+  note = {R package version 1.9.4},
+  url = {https://CRAN.R-project.org/package=fdasrvf},
+}
