### 2021-03-31 0511de2
JOSS paper: minor fixes of references
