+++ paper/paper.bib
@@ -42,7 +42,7 @@
 }
 
 @Manual{R_fda,
-  title = {fda: Functional Data Analysis},
+  title = {{f}da: Functional Data Analysis},
   author = {J. O. Ramsay and Spencer Graves and Giles Hooker},
   year = {2020},
   note = {R package version 5.1.5.1},
@@ -50,7 +50,7 @@
 }
 
 @Manual{R_fdakma,
-  title = {fdakma: Functional Data Analysis: K-Mean Alignment},
+  title = {{f}dakma: Functional Data Analysis: K-Mean Alignment},
   author = {Alice Parodi and Mirco Patriarca and Laura Sangalli and Piercesare Secchi and Simone Vantini and Valeria Vitelli},
   year = {2015},
   note = {R package version 1.2.1},
@@ -58,7 +58,7 @@
 }
 
 @Manual{R_fdasrvf,
-  title = {fdasrvf: Elastic Functional Data Analysis},
+  title = {{f}dasrvf: Elastic Functional Data Analysis},
   author = {J. Derek Tucker},
   year = {2020},
   note = {R package version 1.9.4},
