### 2021-03-31 julia-wrobel/registr#38 @bauer-alex
Thank you very much for the feedback.

Here is what I changed in the package documentation and the paper regarding your recommendations:

1. I adapted Section 2 in the main vignette a bit to make it clearer which distribution can be used in what data situation. Given these changes, I think that the family argument is now documented adequately well.
2. Maybe there was a misunderstanding regarding the use of the analytical version of the gradient. The analytical version of the gradient should always be used when available, since it implements the explicit gradient for the distribution family which otherwise would have to be estimated numerically (i.e., less accurately and less computationally efficient). I see that this might not have been documented perfectly well, leading to some confusion. I changed Section 6 in the main vignette to better explain what we try to say about the difference between the two gradients. Now it should be clear.
3. I added the runtimes for a larger data example in Section 6 in the main vignette.
4. I added such a subsection relating to the dataset at the end of Section 3 in the incomplete curves vignette.
5. Thank you for this valuable point! There is now a new paragraph at the end of Section 1 in the paper (see [main issue](https://github.com/openjournals/joss-reviews/issues/2964)) to list relevant alternative packages.
