### 2021-03-31 #2964 @bauer-alex
@mbobra @qpmnguyen Thanks for the valuable feedback. I now addressed all questions / final recommendations of @qpmnguyen, see [here](https://github.com/julia-wrobel/registr/issues/38). The newly generated pdf in the last comment contains the newest version of the paper.
