### 2021-04-16 0f39a98
allow variationalEM FPCA for Gamma/Poisson data, by performing Gamma/Poisson registration and Gaussian FPCA
