+++ R/bfpca.R
@@ -14,6 +14,8 @@
 #' @importFrom stats rnorm quantile
 #' 
 #' @return An object of class \code{fpca} containing:
+#' \item{fpca_type}{Information that FPCA was performed with the 'variationEM' approach,
+#' in contrast to registr::gfpca_twoStep.}
 #' \item{t_vec}{Time vector over which the mean \code{mu} was evaluated.}
 #' \item{knots}{Cutpoints for B-spline basis used to rebuild \code{alpha}.}
 #' \item{efunctions}{\eqn{D \times npc} matrix of estimated FPC basis functions.}
@@ -235,12 +237,13 @@ bfpca = function(Y, npc = 1, Kt = 8, maxiter = 50, t_min = NULL, t_max = NULL,
   scores     = scores %*% psi_svd$v
   
   ret = list(
+    fpca_type     = "variationalEM",
     t_vec         = t_vec,
     knots         = knots, 
     alpha         = Theta_phi_mean %*% alpha_coefs,
     mu            = Theta_phi_mean %*% alpha_coefs, # return this to be consistent with refund.shiny
     efunctions    = efunctions, 
-    evalues       =  evalues,
+    evalues       = evalues,
     npc           = npc,
     scores        = scores,
     subject_coefs = subject_coef,
