### 2021-04-27 #2964 @mbobra
👋 @bauer-alex We're almost there! Could you please cut a new release and [deposit it in Zenodo](https://guides.github.com/activities/citable-code/)? Please make sure that the author list on the Zenodo deposit matches the author list on the paper. Then I can move forward with accepting the paper!
