### 2021-05-05 afb61a7
removed a deprecated, unused line of code
