### 2021-05-05 8a3cb2c
fixed scores for the orthogonalized solution returned by fpca_gauss and bfpca
