### 2021-05-05 60a242a
fixed scores matrix for npc=1 for fpca_gauss and bfpca
