### 2021-05-05 5864efe
changed verbose from TRUE/FALSE to 0-4 to better control the level of detail
