+++ R/registr.R
@@ -4,7 +4,7 @@ registr_helper = function(obj = NULL, Y = NULL, Kt = 8, Kh = 4, family = "gaussi
                           beta = NULL, t_min = NULL, t_max = NULL, row_obj = NULL,
                           periodic = FALSE, warping = "nonparametric",
                           gamma_scales = NULL, cores = 1L,  subsample = TRUE,
-                          verbose = FALSE,
+                          verbose = 1,
                           ...){ 
   if (!is.null(incompleteness)) {
     if (warping != "nonparametric") {
@@ -85,7 +85,7 @@ registr_helper = function(obj = NULL, Y = NULL, Kt = 8, Kh = 4, family = "gaussi
   } else { # template function = mean of all curves
     
     if (!is.null(Y_template)) {
-      if (verbose) {
+      if (verbose > 2) {
         message("Registr: Extracting Y_template")
       }
       if (!all(c("id", "index", "value") %in% names(Y_template))) {
@@ -99,8 +99,8 @@ registr_helper = function(obj = NULL, Y = NULL, Kt = 8, Kh = 4, family = "gaussi
       mean_dat = Y
     }
     
-    if (verbose) {
-      message("Registr: Getting Knots and basis functions")
+    if (verbose > 2) {
+      message("Registr: Getting nnots and basis functions")
     }
     if (periodic) {
       # if periodic, then we want more global knots, because the resulting object from pbs 
@@ -123,7 +123,7 @@ registr_helper = function(obj = NULL, Y = NULL, Kt = 8, Kh = 4, family = "gaussi
     nrows_basis = nrow(mean_basis)
     # if greater than 10M, subsample
     if (nrows_basis > 10000000 && subsample) {
-      if (verbose) {
+      if (verbose > 2) {
         message("Registr: Running Sub-sampling")
       }       
       uids = unique(mean_dat$id)
@@ -137,7 +137,7 @@ registr_helper = function(obj = NULL, Y = NULL, Kt = 8, Kh = 4, family = "gaussi
       mean_basis = mean_basis[subsampling_index, ]
       rm(subsampling_index)
     }
-    if (verbose) {
+    if (verbose > 2) {
       message("Registr: Running GLM")
     }   
     if (requireNamespace("fastglm", quietly = TRUE)) {
@@ -147,7 +147,7 @@ registr_helper = function(obj = NULL, Y = NULL, Kt = 8, Kh = 4, family = "gaussi
       mean_coefs = coef(mean_coefs)
     } else {
       mean_coefs = coef(glm(mean_dat$value ~ 0 + mean_basis, family = mean_family,
-                            control = list(trace = verbose > 0)))
+                            control = list(trace = verbose > 3)))
     }
     rm(mean_basis)
     rm(mean_dat)
@@ -261,7 +261,9 @@ registr_helper = function(obj = NULL, Y = NULL, Kt = 8, Kh = 4, family = "gaussi
 #' call is parallelized by using \code{parallel::mclapply} (for Unix-based
 #' systems) or \code{parallel::parLapply} (for Windows). Defaults to 1,
 #' no parallelized call.
-#' @param verbose If \code{TRUE}, prints diagnostic messages. Defaults to \code{FALSE}
+#' @param verbose Can be set to integers between 0 and 4 to control the level of
+#' detail of the printed diagnostic messages. Higher numbers lead to more detailed
+#' messages. Defaults to 1.
 #' @param ... additional arguments passed to or from other functions
 #' @param subsample if the number of rows of the data is greater than 
[…]
