### 2021-05-05 1d01e16
minor fix of registr output
