### 2021-05-06 #2964 @kthyng
Hi @bauer-alex, I'll be taking over from here! Can you change the title of your Zenodo archive so the title matches your JOSS paper? This isn't required, but it is a preference to be consistent.
