### 2021-05-06 5b99616
gfpca_covHall: changed default to 'Kc = 10'
